types: PersistentStorage.getItem Promise<T | null>
if the persistent storage is async in nature, we cannot ensure that it will return the value instead of null.

In async behaviour we allow to return `T` or `null`, but in async we assume that it will return a promise of `T`.

We should return `Promise<T | null>` instead.