Allow unserialised data to be written to store
This PR: https://github.com/apollographql/apollo-cache-persist/pull/388
stops the cache being written as anything other than a string.

The CachePersistor accepts an option "serialize: boolean" that can allow
for data to be written to the store without being serialised to a string
first. However with the .toString() coercion, when serialize is true,
all that gets written to the store is [object object], rather than the
object.

This change makes sense, to bring the api inline with the web storage
api. However, it means all data must be stored as JSON, which, when
there is a large data set, is very expensive in memory and cpu.

So this resolves the issue by putting the generic back in, removing the
coercion, and assigning the appropriate generic to each of the storage
wrappers (I think everything is string apart from localForage, but
perhaps other wrappers can support non-string values)